Move get_pub factory function into plugin’s init
Anticipating all public factory functions to be in plugin/__init__
